+++ test/test_ipipneo.py
@@ -80,13 +80,258 @@ class TestIpipNeo(unittest.TestCase):
         personalities = result.get("person").get("result").get("personalities")
         self.assertEqual(len(personalities), 5)
 
+        # 1. O
         self.assertEqual(personalities[0]["openness"].get("O"), 13.262383714475419)
+        self.assertEqual(personalities[0]["openness"].get("score"), "low")
+
+        # 1.1. O. Traits
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][0].get("imagination"), 1
+        )
+        self.assertEqual(personalities[0]["openness"]["traits"][0].get("score"), "low")
+
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][1].get("artistic_interests"),
+            14.832308817371768,
+        )
+        self.assertEqual(personalities[0]["openness"]["traits"][1].get("score"), "low")
+
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][2].get("emotionality"),
+            22.458638690321635,
+        )
+        self.assertEqual(personalities[0]["openness"]["traits"][2].get("score"), "low")
+
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][3].get("adventurousness"),
+            19.18624143043911,
+        )
+        self.assertEqual(personalities[0]["openness"]["traits"][3].get("score"), "low")
+
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][4].get("intellect"), 27.6351617847169
+        )
+        self.assertEqual(personalities[0]["openness"]["traits"][4].get("score"), "low")
+
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][5].get("liberalism"),
+            79.99472618213304,
+        )
+        ##self.assertEqual(personalities[0]["openness"]["traits"][5].get("score"), "high")
+
+        # 2. C
         self.assertEqual(
             personalities[1]["conscientiousness"].get("C"), 18.62283542366208
         )
+        self.assertEqual(personalities[1]["conscientiousness"].get("score"), "low")
+
+        # 2.1. C. Traits
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][0].get("self_efficacy"),
+            67.18081968991703,
+        )
+        # self.assertEqual(
+        #    personalities[1]["conscientiousness"]["traits"][0].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][1].get("orderliness"),
+            18.490917605991513,
+        )
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][1].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][2].get("dutifulness"),
+            28.94752897386354,
+        )
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][2].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][3].get(
+                "achievement_striving"
+            ),
+            9.985264466537785,
+        )
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][3].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][4].get("self_discipline"),
+            17.450496278627668,
+        )
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][4].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[1]["conscientiousness"]["traits"][5].get("cautiousness"),
+            38.524355262005315,
+        )
+        # self.assertEqual(
+        #    personalities[1]["conscientiousness"]["traits"][5].get("score"), "average"
+        # )
+
+        # 3. E
         self.assertEqual(personalities[2]["extraversion"].get("E"), 34.10779114957646)
+        self.assertEqual(personalities[2]["extraversion"].get("score"), "low")
+
+        # 3.1. E. Traits
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][0].get("friendliness"),
+            37.84207122476397,
+        )
+        # self.assertEqual(
+        #    personalities[2]["extraversion"]["traits"][0].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][1].get("gregariousness"),
+            55.874636258650355,
+        )
+        # self.assertEqual(
+        #    personalities[2]["extraversion"]["traits"][1].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][2].get("assertiveness"),
+            10.136565617726376,
+        )
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][2].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][3].get("activity_level"),
+            70.68785093930268,
+        )
+        # self.assertEqual(
+        #    personalities[2]["extraversion"]["traits"][3].get("score"), "high"
+        # )
+
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][4].get("excitement_seeking"),
+            36.25782829248527,
+        )
+        # self.assertEqual(
+        #    personalities[2]["extraversion"]["traits"][4].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][5].get("cheerfulness"),
+            27.560328329400278,
+        )
+        self.assertEqual(
+            personalities[2]["extraversion"]["traits"][5].get("score"), "low"
+        )
+
+        # 4. A
         self.assertEqual(personalities[3]["agreeableness"].get("A"), 1.9889215690788262)
+        self.assertEqual(personalities[3]["agreeableness"].get("score"), "low")
+
+        # 4.1. A. Traits
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][0].get("trust"),
+            51.83212101876228,
+        )
+        # self.assertEqual(
+        #    personalities[3]["agreeableness"]["traits"][0].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][1].get("morality"), 1
+        )
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][1].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][2].get("altruism"),
+            22.561476501041227,
+        )
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][2].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][3].get("cooperation"),
+            1.930347015846948,
+        )
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][3].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][4].get("modesty"),
+            55.67124528467815,
+        )
+        # self.assertEqual(
+        #    personalities[3]["agreeableness"]["traits"][4].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][5].get("sympathy"),
+            21.651269458084244,
+        )
+        self.assertEqual(
+            personalities[3]["agreeableness"]["traits"][5].get("score"), "low"
+        )
+
+        # 5. N
         self.assertEqual(personalities[4]["neuroticism"].get("N"), 58.6632925696303)
+        self.assertEqual(personalities[4]["neuroticism"].get("score"), "high")
+
+        # 5.1. N Traits
+        self.assertEqual(
+            personalities[4]["neuroticism"]["traits"][0].get("anxiety"),
+            55.29153291804727,
+        )
+        # self.assertEqual(
+        #    personalities[4]["neuroticism"]["traits"][0].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[4]["neuroticism"]["traits"][1].get("anger"), 68.23438372956855
+        )
+        # self.assertEqual(
+        #    personalities[4]["neuroticism"]["traits"][1].get("score"), "high"
+        # )
+
+        self.assertEqual(
+            personalities[4]["neuroticism"]["traits"][2].get("depression"),
+            63.91248592040063,
+        )
+        # self.assertEqual(
+        #    personalities[4]["neuroticism"]["traits"][2].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[4]["neuroticism"]["traits"][3].get("self_consciousness"),
+            22.421630518162374,
+        )
+        self.assertEqual(
+            personalities[4]["neuroticism"]["traits"][3].get("score"), "low"
+        )
+
+        self.assertEqual(
+            personalities[4]["neuroticism"]["traits"][4].get("immoderation"),
+            48.1804752623172,
+        )
+        # self.assertEqual(
+        #    personalities[4]["neuroticism"]["traits"][4].get("score"), "average"
+        # )
+
+        self.assertEqual(
+            personalities[4]["neuroticism"]["traits"][5].get("vulnerability"),
+            77.83259998736844,
+        )
+        # self.assertEqual(
+        #    personalities[4]["neuroticism"]["traits"][5].get("score"), "high"
+        # )
 
         #############################################
         # 2. Test with 30 year old woman.
@@ -420,6 +665,115 @@ class TestIpipNeo(unittest.TestCase):
         self.assertEqual(personalities[3]["agreeableness"].get("A"), 6.331832993524202)
         self.assertEqual(personalities[4]["neuroticism"].get("N"), 58.6632925696303)
 
+    def test_facet_scale_level(self) -> None:
+        ipip = IpipNeo(question=120)
+
+        low, high = ipip.get_current_scale_level()
+        self.assertEqual(low, 45)
+        self.assertEqual(high, 55)
+
+        ipip.set_new_facet_level(low_min=40, high_max=90)
+        low, high = ipip.get_current_scale_level()
+        self.assertEqual(low, 40)
+        self.assertEqual(high, 90)
+
+        with self.assertRaises(BaseException) as e:
+            IpipNeo(question=120).set_new_facet_level(low_min="40", high_max=None)
+        self.assertEqual(
+            str(e.exception),
+            "The (low_min) field must be an int!",
+        )
+
+        with self.assertRaises(BaseException) as e:
+            IpipNeo(question=300).set_new_facet_level(low_min=40, high_max="89")
+        self.assertEqual(
+            str(e.exception),
+            "The (high_max) field must be an int!",
+        )
+
+        ipip.set_new_facet_level(low_min=10, high_max=70)
+        low, high = ipip.get_current_scale_level()
+        self.assertEqual(low, 10)
+        self.assertEqual(high, 70)
+
+        ipip = IpipNeo(question=120)
+
+        low, high = ipip.get_current_scale_level()
+        self.assertEqual(low, 45)
+        self.assertEqual(high, 55)
+
+        ipip = IpipNeo(question=300)
+
+        low, high = ipip.get_current_scale_level()
+        self.assertEqual(low, 45)
+        self.assertEqual(high, 55)
+
+    def test_new_norm_scale(self) -> None:
+        with self.assertRaises(BaseException) as e:
+            IpipNeo(question=120).set_new_norm_scale(scale_min="-", scale_max=75)
+        self.assertEqual(
+            str(e.exception),
+            "The (scale_min) field must be an int!",
+        )
+
+        with self.assertRaises(BaseException) as e:
+            IpipNeo(question=120).set_new_norm_scale(scale_min=1, scale_max=None)
+        self.assertEqual(
+            str(e.exception),
+            "The (scale_max) field must be an int!",
+        )
+
+        ipip = IpipNeo(question=120)
+
+        scale_min, scale_max = ipip.get_current_norm()
+        self.assertEqual(scale_min, 32)
+        self.assertEqual(scale_max, 73)
+
+        ipip.set_new_norm_scale(scale_min=27, scale_max=73)
+        scale_min, scale_max = ipip.get_current_norm()
+        self.assertEqual(scale_min, 27)
+        self.assertEqual(scale_max, 73)
+
+        ipip.set_new_norm_scale(scale_min=15, scale_max=60)
+        scale_min, scale_max = ipip.get_current_norm()
+        self.assertEqual(scale_min, 15)
+        self.assertEqual(scale_max, 60)
+
+        ipip = IpipNeo(question=300)
+
+        scale_min, scale_max = ipip.get_current_norm()
+        self.assertEqual(scale_min, 32)
+        self.assertEqual(scale_max, 73)
+
+        ipip = IpipNeo(question=300)
+
+        result = ipip.compute(
+            sex="M", age=40, answers=load_mock_answers_300_1(), compare=False
+        )
+
+        personalities = result.get("person").get("result").get("personalities")
+        self.assertEqual(personalities[0]["openness"].get("O"), 20.534815840469236)
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][0].get("imagination"),
+            1,
+        )
+
+        ipip.set_new_norm_scale(scale_min=25, scale_max=75)
+        scale_min, scale_max = ipip.get_current_norm()
+        self.assertEqual(scale_min, 25)
+        self.assertEqual(scale_max, 75)
+
+        result = ipip.compute(
+            sex="M", age=40, answers=load_mock_answers_300_1(), compare=False
+        )
+
+        personalities = result.get("person").get("result").get("personalities")
+        self.assertEqual(personalities[0]["openness"].get("O"), 20.534815840469236)
+        self.assertEqual(
+            personalities[0]["openness"]["traits"][0].get("imagination"),
+            0.35632569430614325,
+        )
+
     def test_compute_test_custom_300_reversed(self) -> None:
         def d(compare_original: list, compare_reverse: list):
             differences = []
@@ -452,10 +806,10 @@ class TestIpipNeo(unittest.TestCase):
         )
 
         #############################################
-        # 2. Test with 39 year old man.
+        # 2. Test with 40 year old man.
         #############################################
         result = IpipNeo(question=300).compute(
-            sex="M", age=39, answers=load_mock_answers_300_1(), compare=True
+            sex="M", age=40, answers=load_mock_answers_300_1(), compare=True
         )
 
         compare_original = result["person"]["result"]["compare"][
@@ -517,4 +871,22 @@ class TestIpipNeo(unittest.TestCase):
         self.assertEqual(
             len(d(compare_original=compare_original, compare_reverse=compare_reverse)),
             121,
+        )
+
+        #############################################
+        # 6. Test with 40 year old man.
+        #############################################
+        ipip = IpipNeo(question=300)
+        result = ipip.compute(
+            sex="M", age=40, answers=load_mock_answers_300_1(), compare=True
+        )
+
+        compare_original = result["person"]["result"]["compare"][
+            "user_answers_original"
+        ]
+        compare_reverse = result["person"]["result"]["compare"]["user_answers_reversed"]
+
+        self.assertEqual(
+            len(d(compare_original=compare_original, compare_reverse=compare_reverse)),
+            144,
         )