+++ RELEASES.md
@@ -6,6 +6,8 @@ Release **[1.12.0](https://github.com/NeuroQuestAi/five-factor-e/releases/tag/v1
 
 * Revised README files;
 * The minimum and maximum value can be adjusted by the user to normalize it;
+* The user can adjust the (low, average, high) facet levels;
+* Fix (low, average, high) scales;
 * Runs on Python10 or higher.
 
 Release **[1.11.1](https://github.com/NeuroQuestAi/five-factor-e/releases/tag/v1.11.1)**: